concatened favorites and deckCard controllers into user and deck respectively
