+++ src/main/java/com/example/gwent_projet/controller/DeckController.java
@@ -9,6 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.example.gwent_projet.services.DeckService;
+import com.example.gwent_projet.services.dto.deck.DeckCardCreationDTO;
+import com.example.gwent_projet.services.dto.deck.DeckCardDTO;
 import com.example.gwent_projet.services.dto.deck.DeckCreationDTO;
 import com.example.gwent_projet.services.dto.deck.DeckDTO;
 
@@ -57,8 +59,8 @@ public class DeckController {
 	}
 
 	// by ID
-	@GetMapping("/{id}")
-	public ResponseEntity<DeckDTO> getDeckById(@PathVariable("id") Long id) {
+	@GetMapping("/{deck_id}")
+	public ResponseEntity<DeckDTO> getDeckById(@PathVariable("deck_id") Long id) {
 		try { 
 			DeckDTO searchResult = deckService.getDeckById(id);
 			return new ResponseEntity<>(searchResult, HttpStatus.OK);
@@ -70,8 +72,8 @@ public class DeckController {
 
 	// Update deck -----------------------------------------------------------------------------------------------------------------
 	// By ID
-	@PutMapping("/{id}")
-	public ResponseEntity<DeckDTO> updateDeckById(@PathVariable("id") Long id, @RequestBody DeckCreationDTO deck) {
+	@PutMapping("/{deck_id}")
+	public ResponseEntity<DeckDTO> updateDeckById(@PathVariable("deck_id") Long id, @RequestBody DeckCreationDTO deck) {
 		try {
 			// replace user at this ID with the new user
 			DeckDTO updatedDeck = deckService.updateDeck(id, deck);
@@ -85,12 +87,74 @@ public class DeckController {
 
 	// Delete deck -----------------------------------------------------------------------------------------------------------------
 	// By ID
-	@DeleteMapping("/{id}")
-	public ResponseEntity<HttpStatus> deleteDeck(@PathVariable("id") Long id) {
+	@DeleteMapping("/{deck_id}")
+	public ResponseEntity<HttpStatus> deleteDeck(@PathVariable("deck_id") Long id) {
 		try {
 			deckService.deleteDeckById(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		} catch (Exception e) {
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+	// ---- Deck of cards -----
+	
+	// Add card to deck ------------------------------------------------------------------------------------------------------------
+
+	@PostMapping("{deck_id}/card/{card_id}")
+	public ResponseEntity<DeckCardCreationDTO> addCardToDeck(@PathVariable("deck_id") Long deckId, @PathVariable("card_id") Long cardId) {
+		try {
+			DeckCardCreationDTO newDeckCard = deckService.addCardToDeck(deckId, cardId);
+			return new ResponseEntity<>(newDeckCard, HttpStatus.CREATED);
+		} catch (Exception e) {
+			// return null value with ERROR HTTP status
+			return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+	// Get all cards from deck -----------------------------------------------------------------------------------------------------
+
+	@GetMapping("{deck_id}/cards")
+	public ResponseEntity<ArrayList<DeckCardDTO>> getAllCardsInDeck(@PathVariable("deck_id") Long deckId) {
+		try {
+			// new list of cards
+			ArrayList<DeckCardDTO> cards = new ArrayList<DeckCardDTO>();
+			// get all cards from the repository and add them to the list
+			cards.addAll(deckService.getAllCardsInDeck(deckId));
+			if (cards.isEmpty()) {
+				// if no results, return NO CONTENT HTTP status
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			}
+			// return all cards with OK HTTP status
+			return new ResponseEntity<>(cards, HttpStatus.OK);
+		} catch (Exception e) {
+			// return null value with ERROR HTTP status
+			return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+
+	// Delete card from deck -------------------------------------------------------------------------------------------------------
+	// One card
+	@DeleteMapping("/{deck_id}/card/{card_id}")
+	public ResponseEntity<HttpStatus> deleteOneCardFromDeck(@PathVariable("deck_id") Long deckId, @PathVariable("card_id") Long cardId) {
+		try {
+			deckService.deleteOneCardFromDeck(deckId, cardId);
+			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+
+	// All cards
+	@DeleteMapping("/{deck_id}/cards")
+	public ResponseEntity<HttpStatus> deleteAllCardsFromDeck(@PathVariable("deck_id") Long deckId) {
+		try {
+			deckService.deleteAllCardsFromDeck(deckId);
+			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		} catch (Exception e) {
+			e.printStackTrace();
 			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 	}
