+++ src/main/java/com/example/gwent_projet/controller/UserController.java
@@ -9,6 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.example.gwent_projet.services.UserService;
+import com.example.gwent_projet.services.dto.card.CardDTO;
+import com.example.gwent_projet.services.dto.favorites.FavoritesDTO;
 import com.example.gwent_projet.services.dto.user.UserCreationDTO;
 import com.example.gwent_projet.services.dto.user.UserDTO;
 
@@ -67,8 +69,8 @@ public class UserController {
 	}
 
 	// by ID
-	@GetMapping("/{id}")
-	public ResponseEntity<UserDTO> getUserById(@PathVariable("id") Long id) {
+	@GetMapping("/{user_id}")
+	public ResponseEntity<UserDTO> getUserById(@PathVariable("user_id") Long id) {
 		try { 
 			UserDTO searchResult = userService.getUserById(id);
 			return new ResponseEntity<>(searchResult, HttpStatus.OK);
@@ -80,8 +82,8 @@ public class UserController {
 
 	// Update user -----------------------------------------------------------------------------------------------------------------
 	// By ID
-	@PutMapping("/{id}")
-	public ResponseEntity<UserDTO> updateUserById(@PathVariable("id") Long id, @RequestBody UserCreationDTO user) {
+	@PutMapping("/{user_id}")
+	public ResponseEntity<UserDTO> updateUserById(@PathVariable("user_id") Long id, @RequestBody UserCreationDTO user) {
 		try {
 			
 			if (user == null) {
@@ -100,8 +102,8 @@ public class UserController {
 
 	// Delete user -----------------------------------------------------------------------------------------------------------------
 	// By ID
-	@DeleteMapping("/{id}")
-	public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") Long id) {
+	@DeleteMapping("/{user_id}")
+	public ResponseEntity<HttpStatus> deleteUser(@PathVariable("user_id") Long id) {
 		try {
 			userService.deleteUserById(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -110,4 +112,67 @@ public class UserController {
 		}
 	}
 
+	// ---- Favorites ----
+	// Add favorite ----------------------------------------------------------------------------------------------------------------
+
+		@PostMapping("/{user_id}/favorites/{card_id}")
+		public ResponseEntity<FavoritesDTO> addFavorite(@PathVariable("user_id") Long userId, @PathVariable("card_id") Long cardId) {
+			try {
+				FavoritesDTO newFavorite = userService.addFavorite(userId, cardId);
+				return new ResponseEntity<>(newFavorite, HttpStatus.CREATED);
+			} catch (Exception e) {
+				// return null value with ERROR HTTP status
+				return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+
+		// Get all favorites from user -------------------------------------------------------------------------------------------------
+
+		@GetMapping("/{user_id}/favorites")
+		public ResponseEntity<ArrayList<CardDTO>> getAllUserFavorites(@PathVariable("user_id") Long userId) {
+			try {
+				// new list of cards
+				ArrayList<CardDTO> favorites = new ArrayList<CardDTO>();
+				// get all cards from the repository and add them to the list
+				favorites.addAll(userService.getAllUserFavorites(userId));
+				if (favorites.isEmpty()) {
+					// if no results, return NO CONTENT HTTP status
+					return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+				}
+				// return all cards with OK HTTP status
+				return new ResponseEntity<>(favorites, HttpStatus.OK);
+			} catch (Exception e) {
+				// return null value with ERROR HTTP status
+				return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+
+
+		// Delete favorites from user --------------------------------------------------------------------------------------------------
+		// One favorite
+		@DeleteMapping("/{user_id}/favorites/{card_id}")
+		public ResponseEntity<HttpStatus> deleteOneUserFavorite(@PathVariable("user_id") Long userId, @PathVariable("card_id") Long cardId) {
+			try {
+				userService.deleteOneUserFavorite(userId, cardId);
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			} catch (Exception e) {
+				e.printStackTrace();
+				return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+
+		// All favorites
+		@DeleteMapping("/{user_id}/favorites")
+		public ResponseEntity<HttpStatus> deleteAllUserFavorites(@PathVariable("user_id") Long userId) {
+			try {
+				userService.deleteAllUserFavorites(userId);
+				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+			} catch (Exception e) {
+				e.printStackTrace();
+				return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+	
+	
+	
 }
